fix(ctx_read): satisfy LOC gate + regen generated docs
CI failures on the repo-param PR: ctx_read.rs hit the 1500-line gate
after the new tests, and mcp-tools.md was stale against the
ctx_multi_repo description update. Split the repo-param tests into
ctx_read_repo_param_tests.rs (#660 pattern) and regenerated the docs.
